moved Challenge models to challenge package
